+++ Changelog.md
@@ -14,7 +14,8 @@ Bugfixes:
  * Type checker: Proper type checking for bound functions.
  * Type checker: fix crash related to invalid fixed point constants
  * Code generator: expect zero stack increase after ``super`` as an expression.
- * Code Generator: fixed an internal compiler error for ``L.Foo`` for ``enum Foo`` defined in library ``L``.
+ * Code Generator: fix an internal compiler error for ``L.Foo`` for ``enum Foo`` defined in library ``L``.
+ * Code generator: allow inheritance of ``enum`` definitions.
  * Inline assembly: support the ``address`` opcode.
  * Inline assembly: fix parsing of assignment after a label.
  * Inline assembly: external variables of unsupported type (such as ``this``, ``super``, etc.)