+++ Changelog.md
@@ -1,4 +1,4 @@
-### 0.4.3 (unreleased)
+### 0.4.3 (2016-10-25)
 
 Features:
 
@@ -14,9 +14,9 @@ Bugfixes:
  * Type checker: Proper type checking for bound functions.
  * Type checker: fixed crash related to invalid fixed point constants
  * Type checker: fixed crash related to invalid literal numbers.
- * Type Checker: ``super.x`` does not look up ``x`` in the current contract.
+ * Type checker: ``super.x`` does not look up ``x`` in the current contract.
  * Code generator: expect zero stack increase after ``super`` as an expression.
- * Code Generator: fix an internal compiler error for ``L.Foo`` for ``enum Foo`` defined in library ``L``.
+ * Code generator: fix an internal compiler error for ``L.Foo`` for ``enum Foo`` defined in library ``L``.
  * Code generator: allow inheritance of ``enum`` definitions.
  * Inline assembly: support the ``address`` opcode.
  * Inline assembly: fix parsing of assignment after a label.